updated sql stmt

+++ database/daily_refetch.go
@@ -10,7 +10,7 @@ func (db *Db) GetTotalBlocks() (int64, error) {
 // GetMissingBlocks returns an array of missing blocks from one day ago
 func (db *Db) GetMissingBlocks(startHeight, endHeight int64) []int64 {
 	var result []int64
-	stmt := `SELECT generate_series($1 , $2) EXCEPT SELECT height FROM block ORDER BY 1;`
+	stmt := `SELECT generate_series($1::int,$2::int) EXCEPT SELECT height FROM block ORDER BY 1;`
 	err := db.Sqlx.Select(&result, stmt, startHeight, endHeight)
 	if err != nil {
 		return nil
